Moved CGovernanceObject into new files
